MT#59962 UACAuth: add explicit cast to int
Add cast to make it clear that this is an intentional truncation of a
time_t to an int.

Warned-by: Coverity

+++ core/plug-in/uac_auth/UACAuth.cpp
@@ -606,7 +606,7 @@ string UACAuth::calcNonce() {
   HASH RespHash;
 
   time_t now = time(NULL);
-  result = int2hex(now);
+  result = int2hex(static_cast<unsigned int>(now));
   
   MD5Init(&Md5Ctx);
   w_MD5Update(&Md5Ctx, result);
